Added all the scaffolding necessary to support SuperFeedr HTML fragment feeds, including: - Addition of JSoup to the dependencies - Modifications to the SuperFeedrSubscriber to handle the mix of HTML and JSON/XML that you get - Updates to tests to handle the mixed content Upgraded junit to get better assertions, including some changes to old unit tests to fix issues arising.

+++ src/test/java/com/eventswarm/social/events/JsonTweetEventTest.java
@@ -45,7 +45,7 @@ public class JsonTweetEventTest {
     @Test
     public void testGetTweetId() throws Exception {
         JsonTweetEvent instance = (new JsonEventReader()).getFirstEvent(this.getClass().getClassLoader().getResourceAsStream("fixtures/single2hash1mention.json"));
-        assertEquals(384879482202370048L, instance.getTweetId());
+        assertEquals(384879482202370048L, instance.getTweetId().longValue());
     }
 
     @Test
